Update Dockerfile (add README.md) in COPY step to fix GH action error
During the sonar-cli GitHub action test, an error occurred, and the error message is as follows:

#11 5.209 Warning: The current project could not be installed: [Errno 2] No such file or directory: '/code/README.md'
#11 5.209 If you do not want to install the current project use --no-root.
#11 5.209 If you want to use Poetry only for dependency management but not for packaging, you can disable package mode by setting package-mode = false in your pyproject.toml file.
#11 5.209 In a future version of Poetry this warning will become an error!

To resolve this issue, I  add README.md in Docker  COPY step. It's strange that this issue did not occur when testing on your local machine.

This might be due to the difference in Poetry versions between  from machine  and the one installed in the GitHub action setup.